+ HomeViewController, maneja de una forma diferente las gráficas que muestra en su pantalla principal. + PBBar, se creó como clase encargada de hacer gráficas de barras. + PBConstants, se creó para concentrar en un sólo lugar las constantes utilizadas en el PBKit. + PBDataSet, ahora maneja el tamaño de los data sets con un entero no signado de NS. + PBPlot, ahora es heredera de una clase abstracta PBVisualization, además que se removió un método de esta clase y se incluyó en PBUtilities. + PBVisualization y PBXYVisualization se encuentran en desarrollo.

+++ Statwittstics/PBDataSet.h
@@ -7,14 +7,14 @@
 //
 
 #import <Foundation/Foundation.h>
-#import "CorePlot-CocoaTouch.h"
+#import "PBDefines.h"
 
 @interface PBDataSet : NSObject{
     NSString *dataSetTitle;
 
     NSArray *dataPointsX;
     NSArray *dataPointsY;
-    int dataSetLength;
+    NSUInteger dataSetLength;
     
     CPTColor *lineColor;
     CPTColor *fillingColor;
@@ -30,7 +30,7 @@
 //Data sources
 @property (nonatomic, retain) NSArray *dataPointsX;
 @property (nonatomic, retain) NSArray *dataPointsY;
-@property (nonatomic, assign) int dataSetLength;
+@property (nonatomic, assign) NSUInteger dataSetLength;
 
 //Different attributes to set for each plot
 @property (nonatomic, retain) CPTColor *lineColor;
